refactor: apply css modules codemod
