+++ application-templates/starter-typescript/src/components/welcome/welcome.module.css
